add alert and data validation
